Rename from django-prometheus to django-prometheus-metrics
In order to avoid confusion with other packages
