--- django_prometheus/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'django_prometheus.apps.DjangoPrometheusConfig'
