+++ django_prometheus_metrics/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'django_prometheus_metrics.apps.DjangoPrometheusConfig'
