+++ django_prometheus_metrics/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from django_prometheus.views import MetricsView
+from django_prometheus_metrics.views import MetricsView
 
 urlpatterns = [
     url(r'^metrics$', MetricsView.as_view(), name='prometheus-django-metrics')
